Fix markdownlint file
Prettier introduced an error by adding a trailing comma, which is the
default for .jsonc. There essentially is no JSONC specification, and
tools disagree on the exact set of features. VS Code seems to have made
it up without a library or spec; it accepts trailing commas but issues a
warning.

This is such a mess that this commit switches over to YAML.